[JENKINS-15098] - Refactoring Moved base class to another package, modified visibility and javadoc.
Related issues: https://issues.jenkins-ci.org/browse/JENKINS-15098
